+++ packages/ui-components/center-modal.js
@@ -1,4 +1,5 @@
 import React, { Component, Fragment } from 'react';
+import ReactDOM from 'react-dom';
 import cn from './utilities/classnames';
 import { modalWillReceiveProps, } from './utilities/modals';
 const evaluateRenderProp = (prop) => {
@@ -10,7 +11,7 @@ export class CenterModal extends Component {
     }
     render() {
         // Using <> instead of <Fragment> is breaking the linter.
-        return (React.createElement(Fragment, null,
+        return ReactDOM.createPortal(React.createElement(Fragment, null,
             React.createElement("div", { className: cn('center-modal', { 'is-visible': this.props.open, 'is-large': this.props.large }) },
                 this.props.hasX &&
                     (React.createElement("i", { className: "sg-icon sg-icon-x", "data-role": "close-center-modal", onClick: this.props.onClose })),
@@ -18,6 +19,7 @@ export class CenterModal extends Component {
                 evaluateRenderProp(this.props.renderBody),
                 this.props.renderFooter &&
                     (React.createElement("div", { className: "modal-footer" }, evaluateRenderProp(this.props.renderFooter)))),
+<<<<<<< HEAD
             React.createElement("div", { className: cn('modal-mask', { 'is-visible': this.props.open }), onClick: this.props.onClose })));
     }
 }
@@ -26,5 +28,18 @@ CenterModal.defaultProps = {
     hasX: false,
     large: false,
 };
+=======
+            React.createElement("div", { className: cn('modal-mask', { 'is-visible': this.props.open }), onClick: this.props.onClose })), this.props.modalContainer);
+    };
+    CenterModal.defaultProps = {
+        bodyNode: document.body,
+        hasX: false,
+        large: false,
+        modalContainer: document.body,
+    };
+    return CenterModal;
+}(Component));
+export { CenterModal };
+>>>>>>> 3847efcccf5ebca2e4dc3a3fe5bf8592837f3bf0
 export default CenterModal;
 //# sourceMappingURL=center-modal.js.map