+++ packages/ui-components/side-modal.js
@@ -1,9 +1,11 @@
 import React, { Fragment } from 'react';
+import ReactDOM from 'react-dom';
 import cn from './utilities/classnames';
 import { modalWillReceiveProps, } from './utilities/modals';
 export class SideModal extends React.Component {
     componentWillReceiveProps(nextProps) {
         modalWillReceiveProps(nextProps, this.props);
+<<<<<<< HEAD
     }
     render() {
         return (React.createElement(Fragment, null,
@@ -12,5 +14,20 @@ export class SideModal extends React.Component {
     }
 }
 SideModal.defaultProps = { bodyNode: document.body };
+=======
+    };
+    SideModal.prototype.render = function () {
+        return ReactDOM.createPortal(React.createElement(Fragment, null,
+            React.createElement("div", { className: cn('side-modal', { 'is-visible': this.props.isOpen }) }, this.props.children),
+            React.createElement("div", { className: cn('modal-mask', { 'is-visible': this.props.isOpen }), onClick: this.props.onClose })), this.props.modalContainer);
+    };
+    SideModal.defaultProps = {
+        bodyNode: document.body,
+        modalContainer: document.body,
+    };
+    return SideModal;
+}(React.Component));
+export { SideModal };
+>>>>>>> 3847efcccf5ebca2e4dc3a3fe5bf8592837f3bf0
 export default SideModal;
 //# sourceMappingURL=side-modal.js.map